+++ ecd-20200930/alibabacloud_ecd20200930/models/_create_center_policy_request.py
@@ -52,6 +52,7 @@ class CreateCenterPolicyRequest(DaraModel):
         file_transfer_speed: str = None,
         file_transfer_speed_location: str = None,
         gpu_acceleration: str = None,
+        hover_config_msg: str = None,
         html_5file_transfer: str = None,
         internet_communication_protocol: str = None,
         internet_printer: str = None,
@@ -70,6 +71,7 @@ class CreateCenterPolicyRequest(DaraModel):
         mobile_wuying_keeper: str = None,
         mobile_wy_assistant: str = None,
         model_library: str = None,
+        multi_screen: str = None,
         name: str = None,
         net_redirect: str = None,
         net_redirect_rule: List[main_models.CreateCenterPolicyRequestNetRedirectRule] = None,
@@ -96,6 +98,7 @@ class CreateCenterPolicyRequest(DaraModel):
         region_id: str = None,
         remote_coordinate: str = None,
         reset_desktop: str = None,
+        resolution_dpi: int = None,
         resolution_height: int = None,
         resolution_model: str = None,
         resolution_width: int = None,
@@ -290,6 +293,7 @@ class CreateCenterPolicyRequest(DaraModel):
         # *   off: doesn\\"t enable image quality control.
         # *   on: enables image quality control.
         self.gpu_acceleration = gpu_acceleration
+        self.hover_config_msg = hover_config_msg
         # The file transfer feature on the web client.
         # 
         # Valid values:
@@ -359,6 +363,7 @@ class CreateCenterPolicyRequest(DaraModel):
         self.mobile_wuying_keeper = mobile_wuying_keeper
         self.mobile_wy_assistant = mobile_wy_assistant
         self.model_library = model_library
+        self.multi_screen = multi_screen
         # The policy name.
         # 
         # This parameter is required.
@@ -465,6 +470,7 @@ class CreateCenterPolicyRequest(DaraModel):
         self.remote_coordinate = remote_coordinate
         # The computer reset setting.
         self.reset_desktop = reset_desktop
+        self.resolution_dpi = resolution_dpi
         # The resolution height. Unit: pixel. Valid values for cloud applications: 500 to 50000. Valid values for cloud computers: 480 to 4096.
         self.resolution_height = resolution_height
         # The resolution type.
@@ -835,6 +841,9 @@ class CreateCenterPolicyRequest(DaraModel):
         if self.gpu_acceleration is not None:
             result['GpuAcceleration'] = self.gpu_acceleration
 
+        if self.hover_config_msg is not None:
+            result['HoverConfigMsg'] = self.hover_config_msg
+
         if self.html_5file_transfer is not None:
             result['Html5FileTransfer'] = self.html_5file_transfer
 
@@ -888,6 +897,9 @@ class CreateCenterPolicyRequest(DaraModel):
 
         if self.model_library is not None:
             result['ModelLibrary'] = self.model_library
+
+        if self.multi_screen is not None:
+            result['MultiScreen'] = self.multi_screen
 
         if self.name is not None:
             result['Name'] = self.name
@@ -970,6 +982,9 @@ class CreateCenterPolicyRequest(DaraModel):
 
         if self.reset_desktop is not None:
             result['ResetDesktop'] = self.reset_desktop
+
+        if self.resolution_dpi is not None:
+            result['ResolutionDpi'] = self.resolution_dpi
 
         if self.resolution_height is not None:
             result['ResolutionHeight'] = self.resolution_height
@@ -1244,6 +1259,9 @@ class CreateCenterPolicyRequest(DaraModel):
         if m.get('GpuAcceleration') is not None:
             self.gpu_acceleration = m.get('GpuAcceleration')
 
+        if m.get('HoverConfigMsg') is not None:
+            self.hover_config_msg = m.get('HoverConfigMsg')
+
         if m.get('Html5FileTransfer') is not None:
             self.html_5file_transfer = m.get('Html5FileTransfer')
 
@@ -1297,6 +1315,9 @@ class CreateCenterPolicyRequest(DaraModel):
 
         if m.get('ModelLibrary') is not None:
             self.model_library = m.get('ModelLibrary')
+
+        if m.get('MultiScreen') is not None:
+            self.multi_screen = m.get('MultiScreen')
 
         if m.get('Name') is not None:
             self.name = m.get('Name')
@@ -1381,6 +1402,9 @@ class CreateCenterPolicyRequest(DaraModel):
 
         if m.get('ResetDesktop') is not None:
             self.reset_desktop = m.get('ResetDesktop')
+
+        if m.get('ResolutionDpi') is not None:
+            self.resolution_dpi = m.get('ResolutionDpi')
 
         if m.get('ResolutionHeight') is not None:
             self.resolution_height = m.get('ResolutionHeight')
@@ -1874,11 +1898,19 @@ class CreateCenterPolicyRequestClipboardGraineds(DaraModel):
         clipboard_size_unit: str = None,
         clipboard_type: str = None,
         grained_type: str = None,
+        in_clipboard_size: int = None,
+        in_clipboard_size_unit: str = None,
+        out_clipboard_size: int = None,
+        out_clipboard_size_unit: str = None,
     ):
         self.clipboard_size = clipboard_size
         self.clipboard_size_unit = clipboard_size_unit
         self.clipboard_type = clipboard_type
         self.grained_type = grained_type
+        self.in_clipboard_size = in_clipboard_size
+        self.in_clipboard_size_unit = in_clipboard_size_unit
+        self.out_clipboard_size = out_clipboard_size
+        self.out_clipboard_size_unit = out_clipboard_size_unit
 
     def validate(self):
         pass
@@ -1900,6 +1932,18 @@ class CreateCenterPolicyRequestClipboardGraineds(DaraModel):
         if self.grained_type is not None:
             result['GrainedType'] = self.grained_type
 
+        if self.in_clipboard_size is not None:
+            result['InClipboardSize'] = self.in_clipboard_size
+
+        if self.in_clipboard_size_unit is not None:
+            result['InClipboardSizeUnit'] = self.in_clipboard_size_unit
+
+        if self.out_clipboard_size is not None:
+            result['OutClipboardSize'] = self.out_clipboard_size
+
+        if self.out_clipboard_size_unit is not None:
+            result['OutClipboardSizeUnit'] = self.out_clipboard_size_unit
+
         return result
 
     def from_map(self, m: dict = None):
@@ -1915,6 +1959,18 @@ class CreateCenterPolicyRequestClipboardGraineds(DaraModel):
 
         if m.get('GrainedType') is not None:
             self.grained_type = m.get('GrainedType')
+
+        if m.get('InClipboardSize') is not None:
+            self.in_clipboard_size = m.get('InClipboardSize')
+
+        if m.get('InClipboardSizeUnit') is not None:
+            self.in_clipboard_size_unit = m.get('InClipboardSizeUnit')
+
+        if m.get('OutClipboardSize') is not None:
+            self.out_clipboard_size = m.get('OutClipboardSize')
+
+        if m.get('OutClipboardSizeUnit') is not None:
+            self.out_clipboard_size_unit = m.get('OutClipboardSizeUnit')
 
         return self